core-log: improve locked logging primitives
Improve the locked logging by allowing to use pr_lock, pr_unlock and locked
pr_inf/pr_log without a lock flag. Instead use a global shared locking state
that allows nested locking. Add locking to generic pr_* messages. Use a
spinlock using atomic primitives for locking around pr_* locking variables.
Keep track of when a lock was made and which PID made the lock. This allows
us to now check for locked up processes that have exited and not been reaped
or processes that don't yield in a timely way. Allow the locks to detect
these and acquire the lock to allow logging to proceed without a logging
log jam.

For systems that don't provide atomic operations for the spinlock the locking
on logging becomes a no-op.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>