stress-sock: simplyfy the ctrls fetching and free'ing
Use a static buf for the ctrls data so we can populate the ctrls array
with data that does not need strdup'ing, hence de-risking the failures
of low memory strdup's. Also pre-allocate ctrls to be more than enough
space for all the strings and re-allocate it down to the minimal size
once we know how many strings are required in the array.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>