core-helper: add static helper stress_get_meminfo to get memory info
Add a helper to get memory info to remove the number of direct calls
to sysinfo to help a little bit with abstraction. Add a check to
see if swap is enabled to the low memory freeswap check so that we
don't get false positives when swap is low because there is no swap.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>