stress-bad-altstack: for munmap symbol to be fixed up before it is required
The dynamic lib will only fix up a symbol to the munmap system call when it
is needed due to lazy fixup. However, doing this too late in a signal handler
with an illegal alternative stack will cause the dynamic loader issus on
sparc64 causing some very wierd ld warning messages.  Force the use of mumap
on an invalid call just so ensure the symbol is fixed up before we get to
the signal handler.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>