stress-sigfpe: force division by zero for x86, disable alarm for m68K
The optimizer is causing integer divisions to not generate division by
zero traps on x86, so force this by using non-inlined helper functions
at optimization level 0 to try and overcome this.  Also for m68k the
integer divisions by zero sometimes trigger a segfault when a sigalarm
is handled, so disable the alarm and poll the time for the timeout to
avoid this breakage.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>