core-net: add helpers to reserve and release ports
To avoid any port allocation clashes, add helper functions to reserve
and release required ports. If the required port is already in use
then the reservation attempts to find a contiguous free set of ports
and will return the starting port. If it fails, -1 is returned. The
release will release the reserved ports. The port allocation is
implemented using a simple bitmap for ports 0..65535. Ports outside
this range will fail.

Use the API by all the stressors that perform network access with
specific ports.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>