stress-far-branch: re-work mmap handling for 32 bit and improved random mmaps
Re-work the mmap'ing by adding two helpers, the first tries to do 32 bit
mmapings on x86-64 for 32 bit requested addresses otherwise to normal
mmaping. The second helper tries to check of the page is already mmap'd.
If the page cannot be mmap'd then try various random address for a fixed
mapping, falling back to a non-fixed mapping if those fail.

Finally, unroll the indirect jmp call loop to cause more stress on
the branch prediction logic.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>