partial feature testing rework in Makefile.config
There's a new function testing framework (check_func) which gets the job
done with less lines of code because instead of writing a C program to
test the presence of every function, it is now possible to specify the
function's name and the headers it needs. Simultaneously, libbsd is now
acquired through pkg-config, and its "overlay" mode is used which
reduces lines of code even further and adds libobsd support. Many system
functions which weren't previously detected are now detected better,
mostly <wchar.h> and BSD functions.