stress-x86cpuid: more cpuid method for cpu stressor and make new stressor
This stressor is more focused on exercising on instruction on one specific
architecture, so remove it from the more generic cpu stressor. Add metrics
to get some meaningful idea of the cpuid throughput with the 18 different
leaf types.

To improve the optimization, make the stress_x86_cpuid() helper function
an inlined function so it compiles down to 1 op code.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>