stress-str: bundle wcs function args into a struct, improves run time
Passing so many arguments to the str test functions is a inefficient when
some of these args are not used in all the functions. Improve the code by
replacing the args with a single struct that has all the args bundled up.
This also removes the need to voidify unused args to avoid unused arg
warnings. Speeds up the code by a few percent too on x86.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>