core-shared-heap: add strdup for shared description metrics string
The description string is currently a 32 byte array for each instance
and each metric (even unused ones). This is potentially a large
allocation (16 x 32 x number of stressors x number of stressor instances).
We can reduce this by replacing the description array with a pointer
to either a literal string or dup'ing the string using a shared heap.
The dup pre-checks to see if the read-only string is already being used
and will re-use this rather an dup'ing previous existing duplicates.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>