rework stress_strnrnd()
stress_strnrnd() was a function that generated a random string
containing the letters A to Z (in lowercase). It was used in a few
places to generate random bytes rather than a string.

It has been replaced by 2 functions:
stress_rndstr() - Generate a pseudorandom string, now with an expanded
alphabet
stress_rndbuf() - Fill a buffer with pseudorandom bytes