core-mwc: add stress_mwc*modn() functions for modulo'd range
There are numerous occasions in stress-ng where mwc results are being
modulo'd with a value to constrain the range. This however leads to
modulo bias and hence a non-uniform random number distribution. Fix
this by introducing modn helpers that perform the modulo without the
bias.  We scale the desired modulo to large size that fits into the
mwc type size and pick random values until they fit into the range
of 0..scaled_max - 1. Then we can take the modulo without any bias.

See https://research.kudelskisecurity.com/2020/07/28/the-definitive-guide-to-modulo-bias-and-how-to-avoid-it/

Kudos to Guilherme Janczak for spotting and reporting this issue.

Closes https://github.com/ColinIanKing/stress-ng/issues/252

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>