stress-qsort: add an alternative qsort implementation
The libc qsort implementation may change between versions of libc so make
comparisons simpler by adding in a optimized version of the J. L. Bentley
and M. D. McIlroy tuned for 32 bit integers. There is a new --qsort method
option to support the choice, however, libc qsort remains the default.

Also, don't resort data with char sort as this can cause data duplicates
and we need to avoid this. Instread bit flip some bits in the data for a
faster and consistent ordering re-mashing.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>