stress-rawsock: add OOM retries to child processes, back off on ENOBUFS (LP: #1968361)
When running multiple instances of the rawsock stressor we can end up
with ENOBUFS on the send as memory gets low. On ENOBUFS drop the child
priority each time and add a delay to back off and throttle down the
output. Also add OOM respawn on these child processes in case the OOM
killer starts killing the stressor instances.

Fixes: https://bugs.launchpad.net/stress-ng/+bug/1968361

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>