stress-chmod: don't prematurely finish stressors on ENOENT or ENOTDIR
The first stressor to get a SIGALRM tidies up and removes the file that
the multiple chmod instances are chmod'ing, so we get ENOENT/ENOTDIR
errors and the other instances can terminate before their run time is
complete. Since we're still accessing the file via an open file descriptor
it makes sense to keep on exercising on this until all the stressor
instances get a SIGALRM to allow them to run to completion.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>