Skip some testcases if the parameter --oom-avoid is set
We found if a testcase triggers the oom, it could make the whole
system hang with a high rate under the ubuntu core. This issue is
very common on ARM or ARM64 platforms, by now we don't even have one
ARM/ARM64 platform which could run all stress-ng testcases without
system hang under ubuntu core.

The ubuntu core uses squashfs as the mojor filesystem, when the kernel
loads the testcase or libraries from block device to the memory via
squashfs_readahead()/squashfs_readpage(), if an oom occurs, it will
make the system hang by high chance. Please refer to the function
out_of_memory() in the oom_kill.c of linux kernel.

Recently a stress-ng parameter --oom-avoid is added, with this
parameter, some testcases will monitor the size of freemem and avoid
triggering oom, but those testcases still could trigger oom and make
the system hang under the ubuntu core. This is because the criteria of
low memory is different on differnt platforms, it depends on the size
of system memory, the number of cpu, the number of testcase instanace
and ..., it is not easy to set a low memory criteria for all cases.

A temporary solution is to skip those testcases if --oom-avoid is set,
So far according to my test, there are 7 testcases need to be skipped,
if we find more testcases which have this issue in future, we could
adjust the skipping list.

Signed-off-by: Hui Wang <hui.wang@canonical.com>