stress-module: stress finit_module() and delete_module()
There are different ways to test loading modules:

  * modprobe calls
  * kernel module autoloader

To stress tests modprobe we can simply run modprobe in a loop. To stress
test the kernel module autoloader we can stress tests using the upstream
kernel tools/testing/selftests/kmod/kmod.sh. This ends up calling
modprobe in the end, and it has its own caps built-in to self protect
the kernel from too many requests at the same time. Today we cap that
at 50 max concurrent threads at a time.

The userspace modprobe call will also prevent calls if the same module
exists already.

You figured that might be enough to stop usersapce or even the kernel
from doing stupid things, but its not! See the issues reported by
David Hildenbrand on > 400 CPUs with KASAN enabled we get a shortage
of vmap space.

Although I have patches which are in the works which seem to fix this
issue it is pretty hard to reproduce an environment that has 400 CPUs
to stress test what that is like. And although current debugging seems
to indicate that the issue is actually the kernel ACPI driver uses
aliases and each CPU trying to load the same drivers; and although we
have patches in the pipeline to hopefully fix that too [2] -- we still
need a way to stress the hell out of finit_module() because clearly
evidence is showing we could blame either userspace *or* even kernel
for doing stupid things.

This stress-ng modules stressor is designed to help stress test the
finit_module() system call even if the module is already loaded, letting
us tests things which are hard to reproduce as well, things which
otherwise modprobe would pick up on small systems. We allow two methods
to stress test, one with a shared file descriptor, and another where we
call open prior to each finit_module().

Example uses:

sudo ./stress-ng --module 1 --module-name xfs
sudo ./stress-ng --module 1 --module-name xfs --module-sharedfd

The /lib/modules/$(uname -r)/modules.dep file is looked for and
once we find a module which matches your name we use that path.
Since we're not trying to replicate modprobe we don't load any
dependencies for you. Be sure to have them loaded before running
your stress tests.

We don't use libkmod as we don't want to bloat stress-ng, and because
we can provide a simpler parser to get our target module *once*.

[0] https://lore.kernel.org/all/20221013180518.217405-1-david@redhat.com/T/#u
[1] https://lkml.kernel.org/r/bb6e15e0-2831-6352-82c8-92648a29fb0b@redhat.com
[2] https://lkml.kernel.org/r/20230316151036.10181-1-petr.pavlu@suse.com

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>