core-mwc: add modulo mwc functions for max where max is probably a power of 2
The modulo mwc functions where max is a power of 2 are because of the need
to use the % operator. For values of max where there are most likely to be
a power of 2 (but sometimes maybe not) add some helpers to use faster masking
for powers of 2 for an optimization.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>