stress-malloc: add SIGSEGV handler in child
The OOMable child wrapper currently treats SIGSEGVs as a OOM'd process
that needs to be restarted. For the malloc stressor it is useful to
catch SIGSEVs in case the malloc/free stressor has a corrupted heap
and we want to abort prematurely with an error rather than a silent
restart.

Closes https://github.com/ColinIanKing/stress-ng/issues/277

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>