core-resources: ensure resources are marked as free as soon as possible
Currently if stress_resources_allocate bails out (e.g. due to timeouts or
allocation limits hit) then not all of the resource fields are set and the
default allocated memzero'd values are used when free'ing the resources
later. This can be an issue when using resources (such as semaphore ids)
that may be legitimately zero. Fix this by ensuring all the fields are
set to a 'free' value as early as possible before the allocation loop
is executed.

Closes: https://github.com/ColinIanKing/stress-ng/pull/284

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>