stress-io-uring: perform completion at the end of a round of submits
Stash io uring context and keep track of it with user_data. This simplifies
the completion error handling. Perform completions at the end of a round
of submits. Add more duplicated submissions to pad out with more work.
Remove the xattr ops because the original stressor code is not correct
and it does not add much to the I/O stress paths.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>