Add forced_killed_counter to indicate a child has been killed forcefully
There are times where a child process has to be killed forcefully and
in doing so we end up with a state flag indicating the counter was
not updated sanely.  Add a helper to indicate that we know we have force
killed the stressor child so any inconsistent flag state in the counter
is expected. This allows us to differentiate between kills (say by the
OOM killer) and those that are intentional.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>