Add --stderr option, default output now to stdout
Due to a long historical decision, logging of messages went to stdout
because there were originally mainly errors. Overtime all logged messages
ended up being written to stderr which is not correct.  Default all log
output now to stdout and add a --stderr option to allow one to revert
back to the older default.  This effectively makes the --stdout option
a no-op, but lets leave it in to stop any breakage for users that may
still use this.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>