stress-cpu: Deduce waiting time in active cycle from idle time
stress-cpu with cpu-load option to generate partial load works as
following:

First a certain amount of stress code is executed (governed by the value
of cpu-load). When this phase is executed, the active cpu execution time
is computed, using the CLOCK_PROCESS_CPUTIME_ID clock which equates to
scheduler runtime of the current thread.

Second, the idle delay is computed proportionally to the active cpu
execution time:

`delay = (((100 - cpu_load) * (t2 - t1)) / (double)cpu_load)`

This works fine without contention. Let's assume a cpu-load of 40%, and
an active cpu execution of 0.4s, the idle time would be 0.6s and the
toal cycle time 1.0s which indeed corresponds to 40% load.

However, the idle time delay calculation currently ignores the possibility
that the active time may be scheduled out while running, if there are
other tasks in the runq competing for CPU run time.

Let's assume 50% of contention. In this case, the active cpu execution
time of the stress cycle remains 0.4s, but the wall clock time is now
0.8s instead of 0.4s, as for 50% of the time the CPU was executing some
other task. Since the delay computation ignores wall clock, a delay of
0.6s is still computed. Now the total wall clock time for running the
complete cycle is 0.8s + 0.6s = 1.4s, of which 0.4s is active load,
resulting in a load of 28%, below the 40% target.

To fix this problem, deduct the idle time during active cycle from the
idle delay calculation. The idle time during the active cycle is
obtained by subtracting the active time from the wall clock time.

Signed-off-by: Fernand Sieber <sieberf@amazon.com>