core-nt-store.h: fix double precision nt store
The cast of the double to uint64 is clearly wrong as this is turning a
double to a uint64_t. The contents of the double needs storing. Fortunately
using an inlined memcpy and a good optimizer does the necessary trick without
the mempcy. We could do the store on a type-punned value, but this causes
pedantic warnings.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>