stress-stream: enable verification with --verify option
In verification mode we compare previous computes with latest to
see if there any difference. The data is initialized on each verification
loop as we need to compare like with like on the initial conditions. In
non-verify mode, we re-initialize the data every 64 rounds as initialization
is expensive and we need to re-initialize every so often to avoid data
accumulation overflow creep.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>