stress-eigen: add Eigen 2d matrix stress cases
Eigen is a useful test case since it's used by many folk for matrix
computation, so it makes sense to exercise it. Currently this only
builds for gcc + g++ toolchains (not musl-gcc). Add in extra build
tests to see if we have the necessary build environment to build the
C++ eigen matrix ops. If we don't, fall back onto an empty C matrix ops
source.

We also need to link using g++ when building in Eigen code; or fall
back to the normal gcc or other C tool chain if the Eigen library
does not exist or we're not using gcc/g++.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>