stress-ng: add --change-cpu option for force child processes onto another CPU
In some client/server stressor cases it can or may be advantageous on some
systems to force the child process of the client/server pair onto a different
CPU from that of the parent. This new option uses scheduler affinity to try
and move the child process to any CPU other than that of the parent. The
caveat is that the scheduler at any time move the parent to the same CPU
as the child. We could try and prevent that, but that turns the main stress-ng
process management into a mini-scheduler and insanity follows.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>