stress-zlib: fix waitpid() interrupted before child finished processing
Waiting for other processes could be interrupted. In case of `--verify`
it could lead to comparism of unfinished or still processed data. This
happens if the parent process is interrupted while waiting for the child
to complete. In this special case waitpid returns -1 and errno is set to
EINTR. Further informatiuon can be found in `man 3p wait` section
`#RETURN VALUE`.

Fixes: 801faa34ecf5 ("stress-zlib: remove blocking complexity by putting checksum in shared memory")

Co-authored-by: Karsten Graul <kgraul@linux.ibm.com>
Signed-off-by: André Wild <andre.wild3@ibm.com>