stress-ng: speed up child termination by blocking signals and skip excessive free'ing
Child stressor termination can be improved by stopping any signals since we know we're
on a termination path so why service them. Also, don't free up various global heap
resources in the child as the heap will be destroyed on _exit and the time to free
individual objects is unneccessary.  Also, free'ing some resources (such as stressor
objects) is counter productive if we're in a signal context also traversing the
stressor list.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>