core-shim: add a shim wrapper for posix_fallocate for expedited allocations
With some file systems the libc posix_fallocate will write until
the file system is full or the fallocate is complete using repeated
pwrite or write calls and it's not interruptible. On slow file systems
this can take a very long time to complete and to the user it looks like
stress-ng has blocked forever. Add a shim to posix_fallocate that
fallocates in 1MB chunks and also checks for the stress continue flag
and bail out early with a non-POSIXly conforming EINTR return.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>