stress-cache: Add --cache-size option, set cache-size = N [KB], and ignore --cache-level option
This commit adds the --cache-size option to the stress-cache tool.
The --cache-size option allows specifying a custom cache size in
kilobytes (KB) and overrides the --cache-level option.
This enhancement is particularly useful in scenarios where the cache
information is missing in the host machine's cache device tree.
It can be applied in the following situations:

1. When the 'lscpu' command does not display any cache information.
2. When the 'grep . /sys/devices/system/cpu/cpu*/cache/index*/size'
command does not provide valid cache details.
3. When running on an ARM QEMU virtual machine.

By introducing this option, users can continue to perform cache stress
testing by providing a custom cache size even when the cache device
information is inaccessible. This allows for more flexibility and
support in various cache-related testing and simulation scenarios.