stress-hdd: fix verification on partially completed writing due to ENOSPC
Verifiction will randomly fail when ENOSPC occurs and the writes are
not fully complete. The read pass will overcheck the size and will trigger
spurious and incorrect read verification failures. Fix this by keeping
track of the maximum successfully written amount of data and only read
as much as this back to verify.

Fixes: https://github.com/ColinIanKing/stress-ng/issues/305

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>