Manual: Indent stressors to ease man page reading
The current man page file is listing all stressors options in a single
and contiguous list.

It's difficult for readers to estimate what options is tied to which
stressor, especially when names are very similar like :
       --mmapfork N
       --mmapfork-ops N
       --mmapfixed N
       --mmapfixed-mlock
       --mmapfixed-ops N

It's not that easy to see there are two different stressors.

This patch is about identing all options tied to the same stressor under
a title explaining what the stressor is about.

In addition two little changes were done here:
	1/ --wait is moved after vnni stressor to keep alphabetical ordering
        2/ lower-level cache is renamed to last level cache

Signed-off-by: Erwan Velu <e.velu@criteo.com>