core-log: change semantics of pr_lock/unlock
The lock/unlock was intended to stop intermingled writes from
different stressors occurring.  This used a spin lock which causes
a lot of locking contention when running many stressor instances and
this can add large run time delays when stressors rush to write
termination messages at the end of run. A better strategy is to
change the semantics of pr_lock/unlock to pr_block begin/end to
indicate where a block of writes should be written out atomically
with a single write call. The block begin starts buffering messages
and the block end flushes out the buffered messages in one large
write where possible. If buffers cannot be allocated we fall back
to single writes and break the block buffering.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>