core-clocksource: add check to see if HPET clocksource is being used
The HPET clocksource impacts on performance benchmarking on x86 because
it takes thousands of clock cycles to complete and is way more expensive
than reading the TSC. When scaling to many CPUs and performing many
context switches we end up with bottlenecking on the HPET reads and this
impacts heavily on performance benchmarks. Add a pre and post stress
run check and throw a warning message if HPET clocksources are detected.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>