stress-ng: Convert most files to SPDX licenses
SPDX is a standardized identifier for licenses that clearly, and succinctly
identifies licenses.

Many projects such as the Linux kernel are using it, as are many
distributions. It reduces the code size as well.

This patch carefully converts only the files that could be identified
with the tool licensecheck to SPDX. It leaves all copyrights in place
but replaces the license text with the SPDX tag

Signed-off-by: John Kacur <jkacur@redhat.com>