stress-vma: Add a new virtual memory address stressor
This stressor choses a random address space to exercise mmap, munmap,
mlock, munlock and access on 8 pages in the address space. Every 15
seconds it reselects a new virtual memory address to work on. The
dfiferent page operations are run in pthreads to try to force VMA
races when adding/modifying/accessing and removing pages.

Kudos to Vegard Nossum for the original idea and some example code
to base this stressor upon.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>