stress-workload: add a new scheduling workload stressor
The intent of this scheduler is to mimic work items that are
scheduled at random times. By default 100 work items are scheduled for
bursty clustering in a 100,000 microsecond time slice with a 30%
load. This can be configured by an adjustable time slice, a run time
quanta for each work item, a load % (run quanta scaled by load %) and
there are 4 work item scheduling start time scheduling distributions
available.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>