stress-*: improve rate metrics by using harmonic mean
Metrics are either of a rate of something (in which case a harmonic
mean should be used) or of just values (in which case a geometric
mean should be used). Improve the metrics by adding them with
the mean type to use STRESS_GEOMETRIC_MEAN or STRESS_HARMONIC_MEAN.
Update the metrics reporting to compute the appropriate mean based
on the mean type.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>