core-helper: take into consideration NUMA nodes for full cache size
With NUMA systems the total LLC cache normally not a shared cache
amongst all the nodes, the total LLC cache size of a system is
more likely to be the number of NUMA nodes x the per CPU LLC cache
size for a regularly sized cache across all nodes. This is a bit
of an assumption but it is better to over estimate than under
estimate when performing memory bandwitdh benchmarks that need
to ensure write/read buffers are larger than the LLC cache.

Closes https://github.com/ColinIanKing/stress-ng/issues/335

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>