Reduce the file operations if the unused pid can be found within several tries
After studying the case of pthread, we found that when the number of
cores is high, the hot spots are concentrated on native_spin_lock.
Specifically, on the ICX(kernel 6.3.0), with all 160 cores enabled,
native_queued_spin_lock_slowpath dominates over 88% of the total cycles.
As is shown in Fig 1, the call stack analysis indicates the
native_queued_spin_lock_slowpath is frequently invocated for protecting
ref count when multiple threads attempt to open files. Except for the
call stack, most of the other locks are also concentrated in
__x64_sys_openat. By using strace, we found that openat are mostly used
for open the common file “/proc/sys/kernel/pid_max”, which
stress_get_unused_pid_racy() calls for determining the max pid of the
system.

Given the design goal of the pthread stressor, the performance indicator
should reflect the performance of pthread creation. However, the current
focus is on file operations.
Considering that stress_get_unused_pid_racy() is designed to obtain a
non-existent pid, we tried to reduce the file read operations by splitting
the random PID guess into two phases: we guess the unused pid from the
range of 1024-32767 in the first 10 iterations without reading the system
file “/proc/sys/kernel/pid_max”. if the 10 trials fail, we then read the
max_pid from the system file and launch another 10-round guess.

With this revision, the file operations no longer become a bottleneck for
this case. As is shown in Fig 2, the proportion of __x64_sys_openat
decreased from 70.31% to 0.75%.

[Colin Ian King, add more to commit message, remove tab, move buf to keep
 with coding style ]

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>