stress-ng: don't timeout until all stressors are spawned
The manual states:

      -t N, --timeout T
              run each stress test for at least T seconds.

Currently it will terminate prematurely before all the stressors
have been invoked if the timeout has been reached. Fix this to
match the description in the manual. #ifdef the out code using
the STRESS_TERMINATE_PREMATURELY macro, so folks can manually
enable this at build time if they so desire.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>