core-mwc: optimize modulo mwc32 and mwc64 functions
The maximizing of the random value to modulo is currently a shift
loop. Although this is normally a shift,jmp operation it still is
relatively expensive. Replace this with a shift based on the number
of zero leading bits to remove the loop overhead. This requires
adding clz and clzll. Use clzll instead of clzl for 32 bit systems
where long is 32 bits and long long is 64 bits.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>