stress-ng: move ci counter struct into args and args into stat struct
The bogo-op counter shim helpers currently have to dereference the
counter stuct via a pointer and this is an overhead in a heavily
accessed field. Remove the indirection by moving the counter ci
struct into args. This makes args read/writeable so all const
pointer refs to args need to be deconstified.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>