stress-physpage: explicitly use 64bit type for physical addresses
On 32bit systems uintptr_t is a 32bit type, but physical addresses may
require 64bit even on 32bit systems (or 64bit systems running a 32bit
userspace).

This was observed on a ARM AARCH64 system running a 32bit userspace.
Here stress-physpage ended up passing only the lower 32bit of the
physical address to mmap() causing all kinds of crashes accessing the
invalid memory.

To fix this use a 64bit type for the physical address.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>