stress-mmap: reduce affect of mumap on stressor
Currently munmap is page based and is an expensive operation, this
skews the bogo-ops rates for this stressor - the core idea of the
stressor is to hammer mmap calls. When possible, try to munmap
contiguous mapped pages to reduce the munmap call overhead. This
can be an huge performance improvement as this reduces the locking
on the heavily contended spinlock on the lruvec for systems with
large number of cpus and also when using memory cgroups.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>