core-mmap: touch + check just start of each page rather than filling them
stress-mmap just needs to touch and check the start of each page rather
than completely filling them and checking every value. The main overhead
is actually writing to recently mapped pages that cause a page fault
on each new page being written to, so the win here is small but it still
is worth making this optimization. On an Alderlake i9-12900 with 1 CPU
this works out to be ~7.6% faster overall on a 10 second mmap stress
test.

Signed-off-by: Colin Ian King <colin.i.king@intel.com>