stress-mmap: add new option --mmap-write-check for opt-in write + checks
Performing per-page value writes and read checks will force newly memory
mapped pages to be created and faulted-in. While this is a useful test
scenario, it does add a lot of overhead int the mmap stressor, and the
main aim of the stressor is to exercise mapping/remapping operations
rather than a per-page memory santiy check (there are lots of other
stressors such as the vm stressor) to do that already. Make this check
an opt-in with the new --mmap-write-check option.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>