stress-mmap: use non-mod random mwc functions for faster index generation
For a perfectly fair random distribution we should be using the 32/64 bit
versions of stress_mwc*modn_maybe_pwr2(), however these are expensive
operations for cases where n is not a power of 2 which is quite
likely when running on systems with non-power of 2 instances.

Using stress_mwc*() is much faster and is good enough for this
kind of random-ish fast and dirty shuffle operation.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>