core-mwc: optimize and generalize mwc*modn() functions, remove pwr2 helpers
To generate unbiased random values in range 0..max-1 one needs to iteratively
compute new random values until they are in the given range, e.g.

   do {
	val = stress_mwc64();
   } while (val < max);

However, with max being small this ends up with many iterations to
get below the max limit.

One trick is that the following holds true:

       -max % max == ((2^n) - max) % max)
                  == (2^n) % max

..and -max % max is generally very close to the upper bounds of 2^n
for an unsigned integer, so using this as the upper limit of the
do-while loop ends up generally with 1 iteration for most cases
or maybe 2 in most retries, especially when n is large.

Use this optimization for the general case for mwc*modn() functions
and remove the unnecessary specific optimizations for powers of 2
and possible powers of 2 for max.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>