Add shim for dirent d_type
Some operating systems don't provide a dirent d_type field and
some filesystems return d_type as unknown, so we need to use
a stat call on the filename to determine the type. Add a shim wrapper
to return d_type if it exists and is known, or to stat and map
the stat type to the d_type.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>