stress-numa: add --numa-bytes option to specify total memory size
The default case is to use 4M per stressor of memory to exercise
per numa stressor instance. This option allows one to specify the
total memory to be used by all the instances. This will be divided
up by the number of stressors and rounded to the nearest pages, with
at least 1MB per stressor.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>