core-mincore: add missing --page-in OPT_FLAGS_MMAP_MINCORE check
There are a couple of places where the mincore check for the --page-in
option is not being checked in the stress_mincore_touch_pages() and
stress_mincore_touch_pages_generic() functions leading to different
page touching behaviours when --page-in is not enabled. Fix this by
adding in the missing checks.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>