stress-fork: force lazy bind libc functions to be bound before child is run
The child process is forked many times, and each time it has to bind
the libc functions it uses which is a minor overhead. To optimize this
we call these functions before hand in the parent to force the binding
to occur at the earlier point before forking off the children.

Also remove vhanghup since this is not a main focus of the stress test.
Improves performance by ~4.5% on 24 thread Alderlake system.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>