stress-clone: force lazy bind libc functions to be bound before clone is run
The child process is cloned many times, and each time it has to bind
the libc functions it uses which is a minor overhead. To optimize this
we call these functions before hand in the parent to force the binding
to occur at the earlier point before forking off the clones.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>