stress-prio-inv: ensure parent has higher priority than children
For single processor systems we need to ensure the parent is always
runnable by giving it a higher priority than the child processes. Also
don't block the parent waiting for a SIGCHLD or SIGALRM but instead
lighlty poll the keep_stressing flag so we can't block if a child does
not signal it to exit.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>