stress-vm: fix the stress_vm_check argument name
9e3913cd6dac ("stress-vm: speed up prime-0 and prime-1 by reducing total
memory read/writes") introduces the argument name "prime-zero" for the
function stress_vm_check() when it should be referred to as "prime-1",
considering it is invoked from the "prime_one" VM method. This may
cause confusion, particularly when there are bit errors reported.

Let's fix the argument name from "prime-zero" to "prime-one".

Signed-off-by: Amit Singh Tomar <amitsinght@marvell.com>