stress_vm_walking_flush_data: fix buffer overflow on last 7 bytes
The last 7 bytes can overrun, so set the end limit of the buffer being
exercised to buf_len - 7.

Also for systems with mprotect and PROT_NONE, add an extra page to the
end of the buffer and set it to non-read/writable to catch any future
buffer overruns on both read and write access.

Kudos to Gavin Shan for reporting this issue.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>