Remove write potentially outside of buffer
It is possible for reads from the /sys filesystem to rarely return -EBUSY.

When that happens, stress_system_read() will write to a stack location
outside the buffer. Since the buffer is already set to zero at the beginning
with memset will already have cleared the buffer, there's no need for the
else conditional. Remove it.

Closes: https://github.com/ColinIanKing/stress-ng/pull/368

[Colin Ian King] - Add more text to commit message.