stress-ng: exit if deadline and cannot set deadline scheduler
If we set a child to deadline and it fails the admittance test from
sched_setattr() call, then the best thing to do would be to exit at
that point, otherwise the thread ends up trying to use all the cpu
time. This is due to using sched_yield() immediatley returning as
there's no need to wait for the next deadline period.

If the first sched_setattr() fails, then we get an error, but if
if passes then some of the children will just ignore the -EBUSY and
carry on as if nothing had gone wrong.

This is easily testable by turning down the runtime_us parameters on
Linux and then trying to launch too many stress-ng instances. For this
we did:

The second invocation starts but some threads end up being denied
deadline as they've now exceeded our arbitaray limit.

Signed-off-by: Ben Dooks <ben.dooks@codethink.co.uk>