stress-ng: bail out with EXIT_NO_RESOURCE on sched_settings_apply call
Rather than silently failing when trying to set scheduler settings
instead report the error and exit with EXIT_NO_RESOURCE as the resouce
could not be set.  This clears up a couple of issues, such as when the
deadline scheduler is used and -EBUSY is returned when setting the
scheduler or when trying to set the scheduler without the correct
permissions, such as trying to set the FIFO scheduler when not running
as root.

Kudos to Ben Dooks for reporting this issue with a good suggested fix
that helped in reproducing and fixing this issue.

Closes: https://github.com/ColinIanKing/stress-ng/pull/372

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>