stress-fractal: add Mandelbrot and Julia set fractal stressor.
This stressor computes Mandlebrot or Julia sets with a default size
of 1024 x 1024 points and a default of 256 max iterations per point.
The default width, height, number of iterations and method used can
be set using the various fractal stressor options.

The computation of each fractal is split between the N instances of
the stressor with each stressor computing individual rows of the
fractal. A lock protected shared next row helper is used by each
stressor instances to get the row number of the next row to be
computed and the bogo-op counter is bumped once the last row of
each fractal is reached. To ensure computed data stores are not
optimized out, data is stored on a per-row basis, each stressor
instance has 1 a buffer of 1 row.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>